moves home page files into their own directory
