+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Header from './components/header'
-import Body from './components/body'
+import Body from './components/home-page/body'
 import Creator from './components/creator-module/creator'
 // import "./App.scss";
