+++ frontend/src/components/home-page/body.scss
