+++ frontend/src/components/home-page/content-tree.scss
