+++ frontend/src/components/home-page/question-preview.scss
